requirements.txt
python-twitter
markovify
textdistance
